requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
ultralytics==8.3.64
python-multipart==0.0.20
pillow==11.1.0
gunicorn==23.0.0
